Initial setup of engineering playbook
I did this to just setup the basic Zola theme which I took from
https://github.com/codeandmedia/zola_docsascode_theme

The colors still need tweaking, specifically the link colors in dark and
light themes. But I think this is a good theme in terms of navigation of
the content, organization of the content, as well as owning the content
100% rather than it being locked into a system like Confluence. Also
this should be fine as the only people editing or adding to this should
be members of the engineering team and therefore using Git or Markdown
shouldn't be a problem.