Make the link colors better.
I did this so that in dark and light mode the link colors wouldn't be
horrible. I am sure this can be better but I didn't want to spend much
time on dialing in the look of things at first. Just want to get it to
an OK place and then get the content portion going.