Make adidoks theme support disabling blog
I did this so that if someone didn't want the blog navigation they could
easily disable it from the config. I also added a title under the extra
section in the config so that we could effectively have a display title
that is used for the header with this theme. But the full site title
would be used for SEO and everything else.