Setup config & content for adidoks theme
I setup the config & migrated the content over to using the adidoks
theme so that the site would feel a bit better and more up to date. I
also added stuff to the landing page to hopefully try and communicate
some of what it will contain and the value of those things.