Worked on Roles page a bit more
This included organizing things a bit better as well as deciding that we
didn't really want to break down junior, mid, and senior developers as
it is really a spectrum of each of the skills for the most part. So
instead I decided that we should basically create a rubrick that
enumerates the skills so that people can use it to evaluate themselves
against and see what gaps they might have.

I also added the key driving things item to the tech lead role.