Remove 100 lines of change from Journey to small PRs
I am removing this because it was adding more confusion than it was
providing value. The idea with this metric was that it was simply an
indicator that people could use to let them know they need to take a
deeper look at a pull request and if it is logically broken up or not.

People were confused by this and wanted to treat it as a hard threshold.
So I think it is better to just leave it out and have people learn how
to think about PRs in terms of logical chunks.