Make git commit standards use all characteristics
I am modifying the Git Commit standards to make sure that all the
characteristics of commits we need are represented here. The two missing
ones were not-necessarily complete and releasable.

So I rework things to make sure that all of these characteristics are
appropriately defined and make sense within the context of the Git
Commit standards.

I also improved references to our main two articales explaining Git
Patch stack as a methodology. The hope is that this will make it easier
for people starting from the Git Commit standards document to then find
the necessary information to understand the intetion behind Git Patch
Stack and how to do it.