Change Git Commit Message suggested template
I am changing the Git Commit Message suggested template so that it is
less verbose and gets right to the point about what should be there and
not the why of it.

The hope is that this change makes this template and the associated
process feel a bit lighter than having to read the why for it every
time.