Fix incorrect language identifier
I intend for this to correct the language identifier for the verbatim
block so that zola doesn't throw a warning about not being able to find
the language for syntax highlighting.

I am doing this because I don't want the noise when the rendering
happens and because I don't want people to see it as an example and copy
something that is incorrect as an example.

I accomplish this by simplying searching and replacing ```text with ```
in the document.