Modify Git Commit Message Format
I am modifying the git commit message format so that it includes the
git-cl header and possible entries as a starting point. I am also
modifying it so that the template uses comments so that you don't have
to delete the headings in the template and instead can just fill in
under.

I am doing this because I want the git commit message format and
template to make following our processes and standards easier.

To do this I reworked the format a bit to clean it up but also cleaned
up the actual message format template.