Add link to git repo from welcome page
As the text points out, we would like to encourage uptech studio's emplyees to open PRs for this website.
In order to do that, we should make it as easy as possible for them to find the place to do so.

I opened this PR because I followed a link to this page and after reading the welcome page it wasn't clear to me how one would open a PR if they wish to do so.

To achieve that, I added a link directly to the existing relevant text, so no more bulk is added. The link points to the PR page on the git repo.

[changelog]
added: a link to the git repository from the welcome page

ps-id: 903D2637-5023-4BF9-889E-7F7B1B7CA74E