Add LICENSE file & License section to README.md
I am adding the LICENSE file and the License section to the README.md so
that people will know that the content of this repository & site even
visible by the public have all rights reserved by Uptech Works, LLC.

This is important so that we can share our content here but potentially
be able to reuse the content in different mediums in the future.

To accomplish this I simply made sure that the Copyright notice was
present in both the README and the LICENSE file and that they match the
Copyright notice on the website.