Fix "open to Pull Requests" link
This link was previously linking to our LICENSE file. This isn't really
useful at all. So I modified it to go tot he CONTRIBUTING.md instead.

I also removed "Playbook" from the welcome page lead in as it wasn't
needed anymore as the site is much more than the playbook at this point.