Add missing itemListElement to structured data
I intend to fix the structured data with this change so that Google is
happy and can properly parse the structured data.

I am doing this because Google Search Console informed me that the
structured data couldn't be parsed. So I looked into it and found some
examples for breadcrumbs and believe that it is failing because the JSON
key of "itemListElement" was missing even though the items themselves
were present.

The approach that I took was to simply add that JSON key and the
surrounding array square brackets around all the list items. I also made
the schema.org reference https to be consitent with everything else.
Also it seemed to be failing when it wasn't https.