Fix typo in url
My intention was to fix the url which at one point was computed but
doesn't seem to be anymore.

I did this so that the url wouldn't fail to load resource hopefully when
the page loads.

My approach is simply to remove the rest of the computed portion and let
it be a root relative url and see how that goes.