Fix trailing comma issue with structured data
The intension is resolve a trailing comma issue with the structured
data. So I added some conditional checks to address when to include the
trailing comma and when not to include the trailing comma.

I am doing this because I am trying to get the structured data into a
place where it is good and Google is happy.

I am doing this by adding some conditional logic to identify when it is
the lest element in the array and not include the trailing comma in that
case but include it if it is in the middle.