Remove extra heading and clean up typos in agile-development subsection
The intention of this change is to standardize this page with the others
by removing the extra title header. Also cleared up the statement of
what sorts of planning we do, so that the comparison between that and
what 'sprint planning' is can be more easily understood.

The reason for making these changes is to maintain page standards and
succinct clear description of our agile methodologies.

ps-id: 8BC2BA37-7644-4D41-9E40-2709ADEF8137