Update Git Commit Message template
Update Git Commit Message template with refined version that makes it so
we can leave the comments in the file and write the summary above the
first comment and the following sections can be written below their
respective comments. Then when you save and quit it will ignore all the
comments and you will end up with a proper Git Commit Message.